Migrate code from dart:html library to package:web
After upgrade to Flutter 3.29.0 I get a warning:
'dart:html' is deprecated and shouldn't be used. Use package:web and dart:js_interop instead.
Try replacing the use of the deprecated member with the replacement.

And record my gratitude to Dart team for creation of version 1.1.0
of Flutter package web.