Modify text around statement 1
to use explicit question to the reader with a red question mark
right after the end of my statement. To make it very explicit where
the end of the statement is and to make it very explicit that I
expect feedback (agree or don't) .

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 26.0.13 of website https://app.diagrams.net.