Introduce new statement 2 (about meanings of word "згода" in explanatory dictionary)
I think this is important because in future I want to make a
statement about the actual meaning of absence of agreement. To
make that statement I'll need very detailed description of
conditions when a real agreement exists. So, this is the preparation.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 26.0.14 of website https://app.diagrams.net.