Insert new statement 6
to describe a situation when a reader knows about own mistake in
own thoughts about reality. That helped me to realize that just ask
about agreement is not enough. I have to explicitly ask about
presence of mistakes in my statements.

And record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 26.1.1 of website https://app.diagrams.net/.