Insert new statement 12
to prepare explanation of why the articles in explanatory dictionaries
for word "адносіны" is "porridge". Maybe the deeper cause of that was
"porridge" in heads of Belarusians. 🤷 I hope I'll figure out this
eventually.

And record my gratitude to JGraph Ltd and draw.io AG teams for creation
of version 27.0.3 of website https://app.diagrams.net.