Replace statement 27 and remove announcement 17
because I received feedback, which showed weakness of the statement 27.
As the person called that "that is rather terminology statement".
Thus I "dug deeper" and I believe found a better way to describe
requiered aspects of reality.

Record my gratitude to the person, who provided the feedback.

And also record my gratitude to JGraph Ltd and draw.io AG teams
for cretion of version 28.0.3 of website https://app.diagrams.net/.