Move announcement 27 to be right before statement 43
this also caused a change of the announcement number 27 ⟶ 28.
I did that to have the explanations about relationships state near
the statement about relationships state. Before that the announcement
was "too far" from the statement.

And record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 28.1.1 of website https://app.diagrams.net.