Replace statement 16 with "smaller" statements 16, 17, 18, 19, 20
because I received feedback from a couple of people that original
statement 16 is not clear enough to them. And record my gratitude to
the person who reported the issue first. The second person reported
the issue after I decided to split the statement into smaller steps.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 28.2.1 of website https://app.diagrams.net.