Amend statement 25
to make a connection to statement 24 obvious.

And record my gratitude to JGraph Ltd and draw.io AG teams for creation
of version 28.2.7 of website https://app.diagrams.net.

And also record my gratitude to the person whoes feedback helped me
to understand that the connection between statement 25 and 24 may
be not obvious to some people in some cases.