Move statements 36—41 into chapter 2 from chapter 3 as iswithout any modifications
to make it obvious how the chapter 2 was populated. And record my
gratitude to the person who provided feedback on the statements and
asked questions. During several days of discussions I realized that
I have to introduce a separate chapter with appropriate explanations.
And that will also make picture used in comments to statement 35
understood by readers better. Because I'm going to describe rules of
drawing graphs for causes and their effects in the new chapter.